model: Fixing limit borrowing bug in PerOutFaceLimit and PerFibLimit
When the application sends an Interest, router will borrow limit after
forwarding the Interest. If Data gets dropped in the middle, the
application will retransmit the Interest, and router will borrow another
limit. When Data comes back, limit is only returned once. This may cause
the limit to be wasted.

http://redmine.named-data.net/projects/ndnsim refs #1001
GitHub Issues refs #37