model: Adding exclude filter processing in all content store implementations
Although exclude filter is honored now by content stores, forwarding
strategy is completely ignoring them.  That is, if somebody replies to
an interest with a data packet that violates the filter, this data
packet will propagate back to the requester.  This "problem" may or may
not be addressed in future.

Refs #1009 (http://redmine.named-data.net/issues/1009)